requirements.txt
autopep8==2.3.1
click==8.1.7
flake8==7.1.1
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==2.1.5
mccabe==0.7.0
numpy==2.1.0
pandas==2.2.2
pycodestyle==2.12.1
pyflakes==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
six==1.16.0
tzdata==2024.1
yamale==5.2.1
requests==2.32.3
fastapi
prettytable

